Generalize CodeToCpgSuite
