+++ dataflowengineoss-tests/src/test/scala/io/shiftleft/dataflowengineoss/language/DataFlowCodeToCpgSuite.scala
@@ -5,16 +5,16 @@ import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.dataflowengineoss.layers.dataflows.{OssDataFlow, OssDataFlowOptions}
 import io.shiftleft.dataflowengineoss.queryengine.EngineContext
 import io.shiftleft.dataflowengineoss.semanticsloader.{Parser, Semantics}
+import io.shiftleft.fuzzyc2cpg.testfixtures.FuzzyCCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.language.dotextension.ImageViewer
 import io.shiftleft.semanticcpg.layers.{LayerCreatorContext, Scpg}
-import io.shiftleft.fuzzyc2cpg.testfixtures.CodeToCpgSuite
 import overflowdb.traversal.Traversal
 
 import scala.sys.process.Process
 import scala.util.Try
 
-class DataFlowCodeToCpgSuite extends CodeToCpgSuite {
+class DataFlowCodeToCpgSuite extends FuzzyCCodeToCpgSuite {
 
   var semanticsFilename = "dataflowengineoss/src/test/resources/default.semantics"
   var semantics: Semantics = _
