+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/querying/TypeDeclTests.scala
@@ -1,9 +1,9 @@
 package io.shiftleft.fuzzyc2cpg.querying
 
-import io.shiftleft.fuzzyc2cpg.testfixtures.CodeToCpgSuite
+import io.shiftleft.fuzzyc2cpg.testfixtures.FuzzyCCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 
-class TypeDeclTests extends CodeToCpgSuite {
+class TypeDeclTests extends FuzzyCCodeToCpgSuite {
 
   override val code = """
                    | class foo {
