+++ fuzzyc2cpg-tests/src/test/scala/io/shiftleft/fuzzyc2cpg/testfixtures/CodeDirToCpgFixture.scala
@@ -4,7 +4,7 @@ import java.io.File
 
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.semanticcpg.layers.{LayerCreatorContext, Scpg}
-import io.shiftleft.fuzzyc2cpg.testfixtures.LanguageFrontend.FuzzycFrontend
+import io.shiftleft.semanticcpg.testfixtures.LanguageFrontend
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
