+++ queries/src/test/scala/io/shiftleft/queries/FunctionComplexityMetrics.scala
@@ -1,9 +1,9 @@
 package io.shiftleft.queries
 
 import io.shiftleft.semanticcpg.language._
-import io.shiftleft.fuzzyc2cpg.testfixtures.CodeToCpgSuite
+import io.shiftleft.fuzzyc2cpg.testfixtures.FuzzyCCodeToCpgSuite
 
-class FunctionComplexityMetrics extends CodeToCpgSuite {
+class FunctionComplexityMetrics extends FuzzyCCodeToCpgSuite {
 
   override val code = """
     int too_many_params(int a, int b, int c, int d, int e) {
