+++ semanticcpg-tests/src/test/scala/io/shiftleft/semanticcpg/testfixtures/LanguageFrontend.scala
@@ -1,9 +1,8 @@
-package io.shiftleft.fuzzyc2cpg.testfixtures
+package io.shiftleft.semanticcpg.testfixtures
 
 import java.io.File
 
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.fuzzyc2cpg.FuzzyC2Cpg
 
 /**
   * LanguageFrontend encapsulates the logic that translates the source code directory
@@ -23,19 +22,4 @@ abstract class LanguageFrontend {
     * @return CPG representation stored in a file
     */
   def execute(sourceCodeFile: File): Cpg
-}
-
-object LanguageFrontend {
-
-  class FuzzycFrontend extends LanguageFrontend {
-    def execute(sourceCodePath: File): Cpg = {
-      val cpgFile = File.createTempFile("fuzzyc", ".zip")
-      cpgFile.deleteOnExit()
-      val fuzzyc2Cpg = new FuzzyC2Cpg()
-      fuzzyc2Cpg.runAndOutput(Set(sourceCodePath.getAbsolutePath), Set(fileSuffix), Some(cpgFile.getAbsolutePath))
-    }
-    override val fileSuffix: String = ".c"
-  }
-
-  def Fuzzyc: LanguageFrontend = new FuzzycFrontend
 }
